createPaddedArray optimizations for cuda and opencl backend
Before this commit, createPaddedArray was calling createValueArray
to fill the padded regions with a default value which launches a
kernel to do the task at hand. Once that is done, copy kernel was called
upon.

After this commit, createPaddedArray just allocates memory followed by
a kernel copy that takes care of padded regions as well.