install arrayfire cmake configuration and version files
so arrayfire can be used in a cmake project with the usual find_package command

  find_package(ArrayFire REQUIRED)
  include_directories(${ArrayFire_INCLUDE_DIRS})
  tarkget_link_libraries(mytarget ${ArrayFire_LIBRARIES})

When installed in a system location, arrayfire location is discovered automatically
by the project running find_package(ArrayFire REQUIRED). FindArrayFire.cmake
doesn't need to be copied by the project using arrayfire or copied in cmake's
Modules directory.

arrayfire can also be used directly for the build directory.

find_package can search for a specific arrayfire version.

The imported targets automatically deal with MSVC's build configurations.

The installed files are still fully relocatable.

The target files are generated separately for each backend, so the backend
and their config files can be packaged and installed separately. The available
target files are discovered during the configuration of the project using
arrayfire. This should help some linux distro to properly package arrayfire.