Fix identity for char and complex types
The identity kernel defines the ONE and ZERO macros to define those
values to deal with the complex datatypes. This was causing an issue
when the char datatype was used. It was converted into a non-printable
character so the macros were not correctly defined. Also the complex
datatypes did not work because they were converted to string incorrectly