BUGFIX fixed Windows Debug mode issue
copyArray function in all backends was giving undefined reference
errors for specific template specilizations in debug mode. These
specilizations are not used anywhere in code base and are invalid,
hence not added initially. Though the original code was working
fine on non-windows platforms debug mode, on windows it is apparently
causing linking errors. So, to fix this we added the missing
specilizations which throws unsupported error.