Allow building against 64bit index OpenBlas
OpenBLAS can be built with signatures that take a 64bit value as the index type.
When built in this mode blasint is defined to be int64_t, otherwise it is just a
regular int. Since this causes a signature mismatch for those versions of OpenBLAS,
use blasint throughout and provide a fallback typedef for other BLASes.
Also use the OpenBLAS detection to make it use the correct gemm/gemv signature pointer
type. Note that this reverts part of 9be64f05f7236b64949430d4c18ee0421f18bd56, which
I believe was incorrect as it made `BT` pointless. The correct fix would have been
to detect the specific BLAS that was used in the definition below and set the BT to
be `void`.