FEAT Add fallback to CUDA OpenGL Interop
This comes into play when devices are not graphics capable but OpenGL is
available.

On Windows, when all devices are in TCC mode or not connected to display, CUDA
throws an error
`CUDA Error (63): OS call failed or operation not supported on this OS`.

On Linux, if all devices are in TCC mode, the graphics part will still run as
long as OpenGL is available.