BUGFIX: Correct input arrays validation in convolve fns
Earlier to this change, if an n-1 dimensional array and filter were
provided as inputs to af_convolve(n) function, it used to error out
as it expected the nth dimension of arrays to be >1, where n could have
been [1|2|3].

After this change, there is no such restriction. For example,
2d arrays passed onto af_convolve3 should give same output as
passing them to af_convolve2. Have added a test checking the same.