OSX Installer: Get Forge version information from the tag in build_forge.cmake
This is because in the older version, forge version was being computed
from the forge library which is unavailable at CMake time and hence the
variable is empty.

Neither is it possible to fetch the Forge version from forge version
files as the clone happens at build time rather than at CMake time.

Happy to listen to any ideas about how to fix this better.