Remove shared_mutex in favour of thread_local storage
This removes the usage of boost::shared_mutex from graphics,
fft cache management and jit.

Graphcis shall be stated as thread unsafe and hence to be used
from the main thread. Rest of the caches shall use thread_local
in the subsequent commits once an alternative to thread_local
has been implemented for OSX platform.