Workaround enqueueReadBuffer bug on OSX
There seems to be an issue on OSX where the enqueueReadBuffer operation is not
waiting for the previous operations to complete before executing the read
operation. This seems to happen even when passing the CL_TRUE parameter to the
blocking_call parameter of the funciton. I am not able to produce this issue
in a standalone application.