Fix compareArrays test helper to handle NaN
Also fixed the initial value for max op in test helper compareArrays.

This fix to the test helper caused canny image test to fail because
of earlier to the change to compareArrays, the test helper was incorrectly
using FLT_MAX for calculating both minima and maxima of the dataset.
Therefore, corrected the canny image test as well as part of this change.