Change static objects into pointers to fix release issues on Windows
Windows terminates threads before the queue threads and other resources are
released. This causes deadlocks with the condition_variables in the async_queue
objects. This is a bug in Visual Studio/Windows that is documented here:

https://connect.microsoft.com/VisualStudio/feedback/details/747145

This will leak some resources but these resources will be released by the
operating system on exit.