Fix invalid access while exiting with an error in flight
This change fixes an error with the convolution example when
an error was thrown as the arrays were being released. Since these arrays
were created before the global_error_string object they were released after
the error string object was deleted. This caused a segfault sometimes when
the release operation returns an error when the driver's resources have
been released.