Reset the CMAKE_RUNTIME_OUTPUT_DIRECTORY for examples
The runtime output path sets the location where the binaries will be
moved once built. This is done so that the tests and the libraries
are in the same directory on the windows platform. This path is also
set for the examples but some of the example names are similar to
the tests. This causes problems because the examples may overwrite
the test binaries. This commit sets the runtime directory for the
examples.