BUGFIX: correct assets path in examples
ASSETS_DIR was not set to a proper path earlier in image_processing
and machine_learning examples due to which the examples which use
images failed during image loads.

New change adds ASSETS_DIR option to cmake options when the examples
are not built as part of ArrayFire source build.