BUGFIX: Fix OpenCL tests on Windows for CUDA9
The issue fixed in this commit started happening on Windows platform with CUDA9.

Both boost::compute::program_cache and ArrayFire contexts are static
variables. The boost cache was being cleaned up after ArrayFire released
the corresponding OpenCL contexts.

This fix keeps a pointer to the shared_ptr to boost::compute::program_cache
and leaks it on Windows platform to circumvent the problem.