Undef all macros to improve support for unity builds
Unity builds allow you to combine mutiple cpp files in order to reduce the
compilation time by reducing the number of headers that are parsed. This is done
by #including all files into one or more cpp files. This can lead to problems
when the cpp files do not cleanup all the macros and the macro definitions
spill into the next source file. This commit addresses that issue by undefing
all(most?) definitions in cpp files.