Add cmaps missing in arrayfire but in Forge
Sync the `af_colormap` enum with the similar one in Forge.

I have to admit that this PR was written through the GH web interface because I never used arrayfire (yet ;)?), so please correct me if I am wrong! This change looks like a (very) low-hanging fruit, as it is simply a matter of adding a few missing entries compared to the `fg_color_map` enum in the Forge `define.h`, is it not? By doing so, arrayfire users would gain access to the *perceptually uniform* colormaps that were originally designed for Matplotlib a while ago, and have started to spread among other projects. In my opinion, this would be a neat small additional feature to arrayfire, but I may be biased here ;) (regular Matplotlib, to say the least).

Being there, I also added (as comments) a couple information that were in the Forge header file, but not the arrayfire one.