Add (disabled) test for RNG quality
This test repeatedly draws random numbers and generates a histogram from
them. It calculates the chi^2 statistic for the individual step as well
as for the accumulated random numbers. The test fails if two consecutive
statistics are too large or too small. Too large means that the random
numbers are not uniform enough. Too small means that they are "too
uniform", since some amount of random noise is expected.

Additional notes:
* One should also test randn, but that would increase the run-time even
  more.
* The failure conditions are fragile. In principle one should accept a
  larger range of chi^2 values together with more steps, possibly with
  different initial seeds. But this would increase run-time even more.
* As a consequence of the fragile conditions, false positives can occur.
  For example, MERSENNE with double on CPU failed early in some tests.