Remove cmake cmd overrides in examples
There was an infinite loop when building with Microsoft's VCPKG
package manager that was caused because the ArrayFire examples
and vcpkg overloaded the find_package function. This caused a
stack overflow during the CMake configuration step.

Also, removes the find_package override in examples. The target
conflicts are now avoided by prefixing all tests with the string
`test_`.