Make FreeImage a runtime instead of a linktime dependency
This commit will make the FreeImage library a runtime dependency
instead of a link time dependency. This means that the user does
not need to have FreeImage installed on their system if they are
not using the image input and output functions. The FreeImage
library will be loaded at runtime only when the image io functions
are used.

Also includes the DependencyModule class which will make it easier
for other classes to be loaded in a similar way.