Make FreeImage a runtime instead of a linktime dependency
This commit will make the FreeImage library a runtime dependency
instead of a link time dependency. This means that the user does
not need to have FreeImage installed on their system if they are
not using the image input and output functions. The FreeImage
library will be loaded at runtime only when the image io functions
are used.

`DependencyModule` class handles all the symbol management
and it is abstract enough to be used for other libraries too.

Each platform has it's own module_loading file because they were
complete reimplementation of the functions. There are still some
ifdef remaining but they are small enough not to warrent a separate
file.

(cherry picked from commit 659719985d4302f3fd6784d979150f6cacc44173)