Add libdl when building with FreeImage
We are now loading FreeImage at runtime. Because of this we need
to link with the libdl library so that we can call dlopen and dlsym.
This was working on the CI builders because we were linking with
MKL which also requires libdl.

(cherry picked from commit 3e3598b284ee6a21518c383a499d398488f65c47)