Fix host leak when assigning indexed array to another indexed array
Fixes a memory leak caused by assigning from one indexed array to
another indexed array. This was happening because the move constructor
was setting the array_proxy_impl ptr to null and therefore the object
was not getting released. This would have caused a leak in host memory
but the device memory would have been correctly deallocated.