Use sizeof(Param<T>) to determine JIT eval. Add NodeIterator
Uses the the size of the Param<T> object to determine when the JIT node should
be evaluated. This is required because there is a limit to the size of the
parameters you can send to a CUDA kernel.

Also created a NodeIterator which traverses the node tree. This improves the
readability of the code.