Deprecate [u]intl
intl and uintl are not in the global namespace and are not prefixed
by af_ and therefore could conflict with user code. This will deprecate
their usage in user code but maintains those types in the backend.

(cherry picked from commit 65b675b32ba3c8236ea70c0855209e128ee7f751)