FEAT: sparse-sparse add/sub support
CPU and OpenCL backends have support for mul/div but they
are disabled to have feature parity with CUDA which doesn't
have support for mul/div.

The output of sub/div/mul is not guaranteed to have only
non-zero results of the arithmetic operation. The user has
to take care of pruning the zero results from the output.