Load graphics dependencies at runtime
- Forge is no longer a linktime dependency
- glbinding is not a dependency going forward.
- glad GL loader is a compile time dependency, but not required during
link-time or run-time.

If forge and it's dependencies (glfw, fontconfig, freetype) are
available at runtime, graphics functionality will automatically
work as it should, otherwise a runtime exception is thrown.

In order to make this runtime loading of graphics dependencies happen,
all the graphics(forge) C++-API calls have been refactored to use C-API.

An additional, miscallaneous change is that interop manager class and
it's derived classes are cleaned up a bit.