Apply suggestions from clang-tidy tool on array.cpp
Clang-tidy is a tool for static analysis and other quality checks on software.
These changes are some of the suggestions from clang-tidy on the array.cpp file.
This commit does not add the clang-tidy config file because that is still a
work in progress.

(cherry picked from commit 62bf57b2a81975c4d05019e74d63b71ce00b15f5)