Address issues with modified inputs for fftC2R functions.
The fftC2R function was modifying the input array when the fftw library
was used. This was happening because the fftw library does not have
and algorithm that can compute the fft without modifying the input array.

This problem does not seem to be happening on the MKL based fft.