Pass parameters to enqueue by reference instead of by value.
When passing arguments to Queue. We were currently passing all arguments
by value but this was causing issues because the operator Param<T> function
would eval a function as it should. But since we were calling these functions
on a copied Array<T> object, this eval was not propagating to the original
Array<T>.