Fix undefined behavior in CPU RNG for f16 and short
The CPU random number generator was incorrectly calculating the shifts
for the transform function. This logic seemed to be copied from the
uchar transform function. The incorrect behavior only appeared on
Windows and for arrays larger than 256 elements. In this commit I
mask the shifted value before the float factor is calculated.