PERF: Disable events by default. Add enable events flag
This commit disables the use of events but does not remove the
events based code from the memory manager. This is required because
it is not necessary as of now and there is a performance penalty
for using events. This may be enabled in the future once we begin
using streams where events will be necessary.