Refactor memory manager. Remove buffer_info. Simplify allocs
* Removed buffer_info from the allocation calls

The buffer_info object was created so that the event synchonization was
performed in the memAlloc call instead of in the memory manager. It did not have
significant benefit over performing the synchronization within the memory
manager and increased complexity of the API. Now that the events API is
exposed, it shouldn't be necessary to include this.

* Moved memory manager classes into separate files in the common directory
* Use memAlloc instead of bufferAlloc in OpenCL Array<T> initializtion to
  provide type context to the memory manager. Avoid inconsistencies between
  backends.
* Use objects instead of handles within the codebase