Put print_info in the ArrayFire_BINARY_DIR for all Operating systems
The windows runtime binary directory is different than other operating
systems. This causes issues with the print_info command in the
CTEST_CUSTOM_POST_TEST command because it is located in a different
location on windows. This sets the runtime output directory to the
same location on all OSs for print_info