Create a thrust policy to intercept tmp buffer allocations
Thrust uses policies to perform certain operations in the backend. This
commit creates an ArrayFire policy for thrust which intercepts temporary
buffer allocations and frees to the memory manager. It also allows
you to specify the stream of the operation so the older approach to
specify the stream has been updated.