Fix barrierOR fn in confidence connected opencl kernel
Without the extra barrier sync towards end of barrierOR function
after reading the reduction result, the caller's loop if any is
going into infinite loop occasionally which is in turn randoms hangs.

This doesn't seem to be an issue on non-nvidia hardware. Hence, we are
conditionally adding the extra barrier sync conditionally for nvidia
platform.