Delete CLBlast patch and update the tag to include the changes (#2991)
* Fix failure with git apply after make clean

After the clean target is run, the patch on CLBlast is applied again.
this fails because the patch has already been applied. In order to
fix this we need to reset to the original tag and apply again.

* Delete CLBlast patch and update the tag to include the changes

Windows was failing with the previous commit. I updated the tag to
the changes pradeep pushed upstream and deleted the patch