Workaround for new cuSparse API introduced in CUDA patch release
New API of cuSparse was introduced in 10.1.168 for Linux and the older
10.1.105 version doesn't it.

Unfortunately, when the new API was introduced in ArrayFire's code base,
I was testing against versions 10.1.168 or newer and hence didn't realize
that this new API was introduced in a patch/fix release - unconventional.

This change enables the new API only from 10.2.* on Linux since CUDA toolkit
version variable set by CMake doesn't provide patch number.