Fix cmake arguments for external projects for msvc (#3088)
* Fix cmake arguments for external projects for msvc

Without the additional toolset argument being forwarded to msvc
toolchain, cmake/msvc is free to choose a toolset as per their
respective logic. This causes build issues.

Also fixed some conditions that are based on CMakeBuildType variable -
not recommended to use checks based on that variable and often resulted
in issues when used with untested multi-config generators.