Fix bug in getMappedPtr on Arrays that are not ready
Fixes a bug in getMappedPtr where the Array object was not ready
and needed to be evaluated when the map function was called. This
appeared when the LHS or the RHS of the matmul function were
JIT nodes and were sparse Arrays.

(cherry picked from commit 3ff9b242d6f48f088f756b242a364c378cb353e7)