Improve offline build experience for developers
The following common scenarios(majority we think) are covered with this
change.
- Developer has cloud connection always.
- Developer has cloud connection for initial cmake run, but not later.
- Developer has lost cloud connection for a while after the initial
  successful cmake run but regained the connection later.
- Developer is doing an completely disconnected build using the source
  tarball we generate and attach to our release assets.

When the developer wants to do an offline build for a specific commit
other than release tags, they would have to generate the relevant source
tarball themselves. The commands required to do the same can be found
from the following ci workflow file in our repository.

.github/workflows/release_src_artifact.yml

The source tarball generation CI job has also been changed to reflect
the change in external dependencies location.