Check symbols in MKL to enable solve batch functionality
We will first make sure that the getrf_batch_strided function is
available in MKL to determine if the batch functionality can be
used in ArrayFire. If it is available we will define the
AF_USE_MKL_BATCH function to enable the batching functions.

(cherry picked from commit b6680d531ec7ba26e3f3844a05a4654895217488)