Fix reference count if array used in JIT operations.
Previously when an af::array was used in a jit operation and it was backed by a
buffer, a buffer node was created and the internal shared_ptr was stored in the
Array for future use and returned when getNode was called. This increased the
reference count of the internal buffer. This reference count never decreased
because of the internal reference to the shared_ptr.

This commit changes this behavior by createing new buffer nodes for each
call the getNode. We use the new hash function to ensure the equality of
the buffer node when the jit code is generated. This avoids holding the
call_once flag in the buffer object and simplifies the management of
the buffer node objects. Additionally when a jit node goes out of scope
the reference count decrements as expected.