Create a hash function for Node objects for NodeMap
The Node_map_t unordered_map object uses the pointer of the nodes for the key.
This worked because you could previously because the node buffer objects tracked
the buffer object's shared pointer. This required holding an additional
reference to the buffer object when an Array was used in a JIT operation. This
did not leak memory because both the buffer and the node were deleted when the
Array object was destroyed.

This commit creates a new hash function for the node pointers which dereferences
the Node pointers and if they are buffers, it checks the buffer's pointer and
its offset to determine if its unique. This approach allows us to remove the
call_once construct from the setData member function of the buffer node.
You can now create node objects for each invocation getNode function.