Fix nested and duplicate moddims jit issue with the CPU backend
Fix an issue that caused errors with nested moddims caused errors with the CPU
backend. this was caused when the moddims function was called back to back on
the same array.

Another issue that this fixes is when you have the same node which are composed
of moddims arrays in the same jit tree