Prune CUDA static numerical libraries for specifice compute capability
Prune CUDA static libraries so that the binary size of the final executable
is smaller. This commit will run the nvprune utility on some static libraries
(cublasLt, cublas, cusolver, and cusparse) to remove unused architectures
from the binary. The resulting binary is significantly smaller when targeting
a single compute capability.