Make cuSparse a runtime dependency. Optionally allow static linking
This PR adds the ability to load cuSparse at runtime and not at link
time. This allows us to not have cuSparse on the system at startup if
you don't need to use the sparse functionallity in CUDA. Optionally
it also allows you to staticly link against the cuSparse library
if you want to package the library with ArrayFire.