Make constructors that accept simple types explicit
Some of the ArrayFire constructors that accept the dim_t type are not
marked explicit. this allows the initialization of the ArrayFire's array
using integer types. For example

```
af::array a = 5
```

will create an af::array with 5 elements. This is not intended behavior.

I have looked into the ABI for this change and it doesn't seem to
be affected on GCC. I have to still test this on MSVC. This CAN
break some existing code because it does change the API but ArrayFire
was never designed with this code in mind.