Call setDevice on each thread at entry point.
CUDA requires that cudaSetDevice be called in each thread before
any other calls are made to the CUDA API. This is done by default
on the main thread but it is not done on new threads created. This
commit changes the behavior or the af_init function so that it call
the cudaSetDevice when creating a new object in ArrayFire.

This commit also refactors the af_init function so that it calls a lower
overhead init function which initializes the device manager.