Move tile function to common namespace. Avoid calling from detail
This commit moves the implementation of the tile funciton to the common
namespace. This is done because the tile funciton in detail does not perform JIT
optimization. It instead calls the tile kernel directly. This is undesirable
because there are some instances where tile funciton can be performed by
indexing. This commit also updates several calls to tile in the codebase
to use this new version.

It is still fairly easy to call the detail::tile function and we need to address
this at some point. Perhaps it should be deprecated and only called by the
common::tile function. This commit does not address this issue.