Add option to use external dependencies instead of fetchcontent
This commit adds the ability to search for already installed software on the
system instead of downloading the required libraries using fetchcontent. This
allows package managers to select dependencies that are more compatible with the
system than the one targeted by the ArrayFire build system. One disadvantage of
this approach is the increase build failures and version incompatibilities