Avoid overriding `CMAKE_INSTALL_RPATH` on macOS. (#3283)
* Avoid overriding `CMAKE_INSTALL_RPATH` on macOS.

Currently, `InternalUtils.cmake` sets `CMAKE_INSTALL_RPATH` on macOS to
`/opt/arrayfire/lib`. This is not always the install location (e.g. if a
user sets `CMAKE_INSTALL_PREFIX`), nor does it always make sense to only
have a single `LC_RPATH` command inside the libraries on macOS.

In particular, if a user passes `CMAKE_INSTALL_RPATH` from the
command-line on macOS, it would be good to avoid overriding that, since
the user is more likely to supply the correct paths for their system
than keeping a fixed value of `/opt/arrayfire/lib`.

This PR emits a warning if `CMAKE_INSTALL_RPATH` is not set on macOS to
warn the user to set it through the command line.