Avoid installing system forge when AF_INSTALL_STANDALONE not set
The install target was copying the forge library installed on the system. This
is not expected because the install command only copies the artifacts generated
by the project and not libraries installed on the system. We do want system
libraries to be installed when AF_INSTALL_STANDALONE is enabled. This commit
addresses both of these issues.