Initial implementation of OpenCL based JIT for the oneAPI backend
This is the initial implementation of jit based on OpenCL kernels
for the oneAPI backend. This is not a feature complete implementation
and some checks for shape size have been ignored because of the lack
of the memory manager. This implementation also does not fully support
4D tensors which should be implemented later. All jit operations should
be available with this change for most JIT based tests for other
functions.