Remove getData usage in the oneAPI backend and make it private
The getData function should not be used to get the buffer pointer. The get
function should be called instead because the Array could be a node array and in
that case it would be a null pointer. The get function will evaluate the buffer
and then return the resulting array. This was causing crashes after the JIT
changes