Fix race condition in OpenCL kernel
Without the barrier at the end of barrierOR, it is possible for
work-item 0 to start the next loop iteration and update predicates[0]
while other work-items are still inside barrierOR reading `predicates`,
meaning they read the next loop iteration's exit condition. This results
in a divergent loop, where not all work-items reach the same barriers.

A previous fix identified this as a problem only on NVIDIA platforms,
but strictly speaking a barrier is required in all cases to avoid a spec
violation and undefined behaviour.