Fixes to the indexed reduce function for the cpu, opencl and cuda back ends (#3658)
* Fixes to the indexed reduce function for the cpu, opencl and cuda back ends.

These back ends were incorrectly assuming a linear array. In the case of the cuda and opencl back end this was just for the cpu-fallback methods which are used when the total number of elements in the array is less than or equal to 4096.

* Ensure array is evaluated before reducing

Added an eval() to the input array on the CPU back end for the ireduce method to ensure that the array has been evaluated before reducing.